src/dashboard/requirements.txt
pandas
scikit-learn
mlflow
streamlit
uvicorn
psycopg2-binary
Flask
joblib
Werkzeug
python-dotenv
seaborn 